+++ R/use_skill_document.R
@@ -27,17 +27,12 @@ use_skill_document <- function(
 
   shared_params_path <- usethis::proj_path("R/aaa-shared_params.R")
   if (!fs::file_exists(shared_params_path)) {
-    .use_template(
-      "aaa-shared_params.R",
-      "R/aaa-shared_params.R",
-      list(),
-      open = FALSE
-    )
+    .use_template("aaa-shared_params.R", "R/aaa-shared_params.R")
     cli::cli_inform(c(
       "{.file R/aaa-shared_params.R} created.",
       "i" = paste(
         "Add shared parameters to it as your package grows.",
-        "Functions can inherit them with {.code @inheritParams .shared-params}."
+        "Functions can inherit shared parameters with {.code @inheritParams .shared-params}."
       )
     ))
   }